refactor(test): migrate telegram.ts helpers to three-tier test architecture
Move 11 seeder functions + signTestConnectParams to db-test-seeders/telegram.ts
with @why-db-direct JSDoc.

Move 5 assertion functions to db-test-assertions/telegram.ts.

Re-exports in api-test-helpers/telegram.ts preserve backward compatibility.
No test files modified.

Closes #9352
